refactor(pipeline): rename pipeline creation file from dependency check to `create-tests`
